add simple_humanoid example
Copied and adapted from https://github.com/laas/simple_humanoid

The model generated from this data builds, is not totally identical to the
example from metapod:

 * the one from metapod has all its axes oriented along the X-axis
   this one has the oriented along X, Y or Z. As a consequence, the
   generated model uses a combination of REVOLUTE_AXIS_X and
   REVOLUTE_AXIS_ANY.

 * the joint::positionInConf from the metapod model does not follow
   the ordering of the joint. As a consequence, the two models do
   not load the configuration vector the way.

Because of the second problem, the tests do not pass.

Maybe it could work in the zero configuration.

A solution would be to let the user customize each joint positionInConf.